+++ pages/api/books/[bookId].ts
@@ -1,5 +1,6 @@
 import type {NextApiRequest, NextApiResponse} from "next"
 import {services} from "../../../service/Services"
+import {BookInfoUpdate} from "../../../model/Book"
 
 function getParams(req: NextApiRequest) {
   const {bookId} = req.query
@@ -11,6 +12,7 @@ function getParams(req: NextApiRequest) {
 interface Body {
   ocr: boolean | undefined
   currentPage: number | undefined
+  info: BookInfoUpdate | undefined
 }
 
 export default async function handler(
@@ -29,6 +31,9 @@ export default async function handler(
     await services.bookService.ocrBook(bookId)
   } else if (body.currentPage) {
     await services.bookService.updateBookProgress(bookId, body.currentPage)
+    res.status(200).end()
+  } else if (body.info) {
+    await services.bookService.updateBookInfo(bookId, body.info)
     res.status(200).end()
   } else {
     res.status(400).end()