+++ util/Url.ts
@@ -43,3 +43,13 @@ export function isValidWebSocketUrl(value: string) {
   }
   return url.protocol === "ws:" || url.protocol === "wss:"
 }
+
+export function isValidWebUrl(value: string) {
+  let url
+  try {
+    url = new URL(value)
+  } catch (_) {
+    return false
+  }
+  return url.protocol === "http:" || url.protocol === "https:"
+}