+++ CHANGES.md
@@ -1,12 +1,13 @@
 #### Version 1.2
-- Fixed bug where resetting the timer did not reset page read mark
+- Improved home screen now supports editing books description, notes and source. You can also pin and archive books.
 - Minimum OCR confidence can now be changed on the fly in the reader settings
+- Fixed bug where resetting the timer did not reset page read mark
 
 #### Version 1.1
 - Improved overlay now supports automatic font sizing and better text direction detection
 - Added reading timer tracking elapsed time and reading speed
-- Fixed bug where text selection is not cleared when switching pages
 - Text hooker WebSocket URL can now be changed in settings
+- Fixed bug where text selection is not cleared when switching pages
 
 #### Version 1.0
 - Initial release